Fixed memory problem in network reading code
Graphs were being converted to dense matrices, which caused problems
with large graphs.  There is a functionality change that does not need
to be documented because the docs were not explicit before.  But now if
there are two entries for a single pair in the form of (node1, node2,
value1) and (node2, node1, value2), these will be considered parallel
resistors and their conductances will be summed.